Refactor: Restructure test files and fix potential errors
- I moved test files from `src/test` to be co-located with their respective components in `src/components/` and `src/`.
- I renamed test files to `*.test.tsx` convention.
- I updated import paths in test files to reflect new locations.
- I updated `vitest.config.ts` to discover tests in new locations and use the moved `setupTests.ts`.
- I removed the `src/test` directory.

Fix: Improve test selector robustness in calculator tests

- I added `data-testid` attributes to input fields in `BonusCalculator.tsx` and `CTSCalculator.tsx`.
- I updated corresponding tests in `BonusCalculator.test.tsx` and `CTSCalculator.test.tsx` to use these `data-testid` attributes for selecting elements. This makes the tests less brittle and less reliant on display values or placeholder text, potentially fixing four previously reported errors.
