+++ src/App.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen, fireEvent } from '@testing-library/react';
 import { describe, it, expect } from 'vitest';
-import App from '../App';
+import App from './App';
 
 describe('App', () => {
   it('renders main header', () => {
