+++ src/components/ProfitSharingCalculator.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen, fireEvent } from '@testing-library/react';
 import { describe, it, expect, vi } from 'vitest';
-import ProfitSharingCalculator from '../components/ProfitSharingCalculator';
+import ProfitSharingCalculator from './ProfitSharingCalculator';
 
 const mockProps = {
   sueldoBasico: '3000',
